Split Command and Hypervisor tests
